requirements.txt
numpy
gym
matplotlib
Pandas
Pillow